DS-489: Rename Checkbox Group Component
