+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/checkbox-group/checkbox-group-option/checkbox-group-option.component.ts
@@ -20,13 +20,13 @@ import {
 import { FormControl, FormGroup } from '@angular/forms';
 
 @Component({
-  selector: 'fudis-checkbox',
-  templateUrl: './checkbox.component.html',
-  styleUrls: ['./checkbox.component.scss'],
+  selector: 'fudis-checkbox-group-option',
+  templateUrl: './checkbox-group-option.component.html',
+  styleUrls: ['./checkbox-group-option.component.scss'],
   encapsulation: ViewEncapsulation.None,
   standalone: false,
 })
-export class CheckboxComponent implements OnInit, OnDestroy {
+export class CheckboxGroupOptionComponent implements OnInit, OnDestroy {
   constructor(
     private _idService: FudisIdService,
     @Host() protected _checkboxGroup: CheckboxGroupComponent<object>,
