+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/checkbox-group/checkbox-group.mdx
@@ -1,6 +1,6 @@
 import { ArgTypes, Meta, Canvas } from "@storybook/blocks";
 import { CheckboxGroupComponent } from "./checkbox-group.component";
-import { CheckboxComponent } from "./checkbox/checkbox.component";
+import { CheckboxGroupOptionComponent } from "./checkbox-group-option/checkbox-group-option.component";
 import * as CheckboxGroupStories from "./checkbox-group.stories";
 import { checkboxGroupExclude } from "../../../utilities/storybook";
 
@@ -199,4 +199,4 @@ Popover can be included in the checkbox group fieldset with properties from [Pop
 
 ## Properties of Checkbox Component
 
-<ArgTypes of={CheckboxComponent} />
+<ArgTypes of={CheckboxGroupOptionComponent} />
