+++ ngx-fudis/projects/ngx-fudis/src/lib/ngx-fudis.module.ts
@@ -25,8 +25,9 @@ import { BreadcrumbsComponent } from './components/breadcrumbs/breadcrumbs.compo
 import { BreadcrumbsItemComponent } from './components/breadcrumbs/breadcrumbs-item/breadcrumbs-item.component';
 import { BodyTextComponent } from './components/typography/body-text/body-text.component';
 import { ButtonComponent } from './components/button/button.component';
-import { CheckboxComponent } from './components/form/checkbox-group/checkbox/checkbox.component';
+import { CheckboxComponent } from './components/form/checkbox/checkbox.component';
 import { CheckboxGroupComponent } from './components/form/checkbox-group/checkbox-group.component';
+import { CheckboxGroupOptionComponent } from './components/form/checkbox-group/checkbox-group-option/checkbox-group-option.component';
 import { DatepickerComponent } from './components/form/date/datepicker/datepicker.component';
 import { DateRangeComponent } from './components/form/date/date-range/date-range.component';
 import { DescriptionListComponent } from './components/description-list/description-list.component';
@@ -155,8 +156,8 @@ import { PopoverDirective } from './directives/popover/popover.directive';
     BreadcrumbsComponent,
     BreadcrumbsItemComponent,
     ButtonComponent,
-    CheckboxComponent,
     CheckboxGroupComponent,
+    CheckboxGroupOptionComponent,
     ControlComponentBaseDirective,
     DatepickerComponent,
     DateRangeComponent,
@@ -241,6 +242,7 @@ import { PopoverDirective } from './directives/popover/popover.directive';
    * E.g. components from Angular Material or other Angular tools
    */
   imports: [
+    CheckboxComponent,
     CommonModule,
     FormsModule,
     LayoutModule,
@@ -272,6 +274,7 @@ import { PopoverDirective } from './directives/popover/popover.directive';
     ButtonComponent,
     CheckboxComponent,
     CheckboxGroupComponent,
+    CheckboxGroupOptionComponent,
     DatepickerComponent,
     DateRangeComponent,
     DateStartDirective,
